LayoutTests/media/video-single-valid-source-expected.txt

Test that a single valid <source> element loads correctly


EVENT(loadedmetadata)
END OF TEST
